DEPR/API: Non-ns precision in Index constructors
This deprecates passing dtypes without a precision to DatetimeIndex
and TimedeltaIndex

```python
In [2]: pd.DatetimeIndex(['2000'], dtype='datetime64')
/Users/taugspurger/.virtualenvs/pandas-dev/bin/ipython:1: FutureWarning: Passing in 'datetime64' dtype with no precision is deprecated
and will raise in a future version. Please pass in
'datetime64[ns]' instead.
  #!/Users/taugspurger/Envs/pandas-dev/bin/python3
Out[2]: DatetimeIndex(['2000-01-01'], dtype='datetime64[ns]', freq=None)
```

Previously, we ignored the precision, so that things like

```
In [3]: pd.DatetimeIndex(['2000'], dtype='datetime64[us]')
Out[3]: DatetimeIndex(['2000-01-01'], dtype='datetime64[ns]', freq=None)
```

worked. That is deprecated as well.

Closes https://github.com/pandas-dev/pandas/issues/24739
Closes https://github.com/pandas-dev/pandas/issues/24753

+++ pandas/core/indexes/base.py
@@ -311,7 +311,8 @@ class Index(IndexOpsMixin, PandasObject):
         elif (is_timedelta64_dtype(data) or
               (dtype is not None and is_timedelta64_dtype(dtype))):
             from pandas import TimedeltaIndex
-            result = TimedeltaIndex(data, copy=copy, name=name, **kwargs)
+            result = TimedeltaIndex(data, copy=copy, name=name, dtype=dtype,
+                                    **kwargs)
             if dtype is not None and _o_dtype == dtype:
                 return Index(result.to_pytimedelta(), dtype=_o_dtype)
             else:
